Improve code-sharing in e2e tests
