+++ test/editorHelper.ts
@@ -1,0 +1,12 @@
+import * as vscode from 'vscode';
+
+export async function showFile(docUri: vscode.Uri) {
+  const doc = await vscode.workspace.openTextDocument(docUri);
+  return await vscode.window.showTextDocument(doc);
+}
+
+export async function setEditorContent(editor: vscode.TextEditor, content: string): Promise<boolean> {
+  const doc = editor.document;
+  const all = new vscode.Range(doc.positionAt(0), doc.positionAt(doc.getText().length));
+  return editor.edit(eb => eb.replace(all, content));
+}
