+++ test/interpolation/features/definition/basic.test.ts
@@ -1,7 +1,8 @@
 import * as assert from 'assert';
 import * as vscode from 'vscode';
-import { showFile } from '../../helper';
-import { getDocUri, position, sameLineLocation } from '../../util';
+import { showFile } from '../../../editorHelper';
+import { position, sameLineLocation } from '../../../util';
+import { getDocUri } from '../../path';
 
 describe('Should find definition', () => {
   const docUri = getDocUri('definition/Basic.vue');
