+++ test/lsp/features/diagnostics/noScriptRegion.test.ts
@@ -1,5 +1,5 @@
-import { getDocUri } from '../../util';
-import { testNoDiagnostics } from './helper';
+import { testNoDiagnostics } from '../../../diagnosticHelper';
+import { getDocUri } from '../../path';
 
 describe('Should find no error when <script> is not present', () => {
   const childUri = getDocUri('diagnostics/noScriptRegion/Child.vue');
