+++ test/lsp/features/documentColor/basic.test.ts
@@ -1,7 +1,8 @@
 import * as assert from 'assert';
 import * as vscode from 'vscode';
-import { showFile } from '../../helper';
-import { getDocUri, position, sameLineRange } from '../../util';
+import { showFile } from '../../../editorHelper';
+import { position, sameLineRange } from '../../../util';
+import { getDocUri } from '../../path';
 
 describe('Should do documentColor', () => {
   const docUri = getDocUri('documentColor/Basic.vue');
