+++ test/util.ts
@@ -1,5 +1,4 @@
 import * as vscode from 'vscode';
-import * as path from 'path';
 
 export function position(line: number, char: number) {
   return new vscode.Position(line, char);
@@ -16,10 +15,6 @@ export function location(uri: vscode.Uri, startLine: number, startChar: number, 
 export function sameLineLocation(uri: vscode.Uri, line: number, startChar: number, endChar: number) {
   return new vscode.Location(uri, sameLineRange(line, startChar, endChar));
 }
-
-export const getDocPath = (p: string) => {
-  return path.resolve(__dirname, '../../../test/vue3/fixture', p);
-};
-export const getDocUri = (p: string) => {
-  return vscode.Uri.file(getDocPath(p));
-};
+export function sleep(ms: number) {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
